+++ src/EventHub/EventHub/ChangeLog.md
@@ -18,6 +18,8 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Added new Cluster cmdlets - New-AzEventHubCluster, Set-AzEventHubCluster, Get-AzEventHubCluster, Remove-AzEventHubCluster, Get-AzEventHubClustersAvailableRegions
+* Fix for issue 10722 : Fix for assigning only 'Listen' to AuthorizationRule rights.
 
 ## Version 1.5.0
 * Added Managed Identity parameters to `New-AzEventHubNamespace` and `Set-AzEventHubNamespace` cmdlets